Split diagnostic core
